+++ test/containers/rootfs/should-display-the-help-text-if-no-arguments-are-passed.sh
@@ -1,0 +1,11 @@
+#!/bin/sh
+
+set +e
+
+HELP="$(../../../dist/ntenvs c r)"
+EXPECT="Manages containers as simple folders. (Under construction..)"
+
+if ! echo "${HELP}" | grep -q "${EXPECT}"; then
+  printf "Expected:\\n\\n%b\\n\\nActual:\\n\\n%b" "${EXPECT}" "${HELP}"
+  exit 1
+fi