+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `containers disk-image` subcommand
 - `containers rootfs` subcommand
 - `languages` subcommand
+- `projects` subcommand
 - `templates` subcommand
 - Travis CI integration
 - Better Code Hub integration
+- Codacy integration